Choosing candidate by optimising transformation matrix
